fix geo_modif persistance #14313

+++ src/main/java/fr/microtec/geo2/persistance/repository/common/GeoModifCorpsRepository.java
@@ -1,10 +1,12 @@
 package fr.microtec.geo2.persistance.repository.common;
 
+import java.math.BigDecimal;
+
 import org.springframework.stereotype.Repository;
 
 import fr.microtec.geo2.persistance.entity.common.GeoModificationCorps;
 import fr.microtec.geo2.persistance.repository.GeoRepository;
 
 @Repository
-public interface GeoModifCorpsRepository extends GeoRepository<GeoModificationCorps, Integer> {
+public interface GeoModifCorpsRepository extends GeoRepository<GeoModificationCorps, BigDecimal> {
 }
